Added Major adjustments

+++ pages/api/products/[id].js
@@ -3,9 +3,12 @@ import Product from "../../../models/Product";
 
 export default async function handler(req, res) {
   const {
+    cookies,
     method,
     query: { id },
   } = req;
+
+  const token = cookies.token;
 
   await dbConnect();
 
@@ -19,6 +22,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true,
@@ -30,6 +36,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authenticated!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("Product deleted successfully!");
